Add files via upload
Connected fa icon CDN and added the sun and moon icons to the top of the page. These are static for now but represent the idea I have for light/dark mode. I also implemented a dark/light mode color palette on the CSS by declaring variables for both light and dark. Currently sitting in "dark mode" until I add the JS functionality.